% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaShiny-package.R
\docType{package}
\name{gammaShiny-package}
\alias{gammaShiny-package}
\alias{_PACKAGE}
\alias{gammaShiny}
\title{gammaShiny: 'Shiny' Applications for the R Package 'gamma'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of 'shiny' applications for the R package 'gamma'.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab gammaShiny \cr
 \strong{Type:} \tab Package \cr
 \strong{Version:} \tab 0.2.0 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{DOI:} \tab \doi{10.5281/zenodo.4139005}
}
}
\note{
This work received a state financial support managed by the Agence Nationale
de la Recherche (France) throught the program *Investissements d'avenir*
(ref. \href{https://lascarbx.labex.u-bordeaux.fr}{ANR-10-LABX-52} and
\href{https://amidex.univ-amu.fr/}{11-IDEX-0001}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/crp2a/gammaShiny}
  \item Report bugs at \url{https://github.com/crp2a/gammaShiny/issues}
}

}
\author{
**Full list of authors and contributors** (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab *Université Bordeaux Montaigne, France* \cr
 Brice Lebrun \tab *Université Bordeaux Montaigne, France*
}

**Package maintainer**

Nicolas Frerebeau\cr
\email{services-archeosciences@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
