% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny App}
\usage{
launch_app(app = c("doserate", "calibration"))
}
\arguments{
\item{app}{A \code{\link{character}} string specifying the Shiny application
to launch. It must be one of \code{"doserate"} or \code{"calibration"}
(see details).}
}
\value{
A \pkg{shiny} application object.
}
\description{
A wrapper for \code{\link[shiny]{shinyAppDir}}.
}
\details{
\tabular{ll}{
  **Application name** \tab  **Keyword** \cr
  Dose rate estimation \tab \code{doserate} \cr
  Calibration curve builder \tab \code{calibration}
 }
}
\examples{
\dontrun{
launch_app("doserate")
launch_app("calibration")
}
}
\author{
N. Frerebeau
}
\concept{shiny}
