% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_about.R
\name{mod_ui_about}
\alias{mod_ui_about}
\title{About}
\usage{
mod_ui_about(id)
}
\arguments{
\item{id}{TODO.}
}
\description{
About
}
\seealso{
Other module: 
\code{\link{mod_ui_energy}()},
\code{\link{mod_ui_import}()},
\code{\link{mod_ui_settings}()}
}
\concept{module}
