% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_energy.R
\name{mod_ui_energy}
\alias{mod_ui_energy}
\title{Energy Scale Calibration}
\usage{
mod_ui_energy(id)
}
\arguments{
\item{id}{TODO.}
}
\description{
Energy Scale Calibration
}
\seealso{
Other module: 
\code{\link{mod_ui_about}()},
\code{\link{mod_ui_import}()},
\code{\link{mod_ui_settings}()}
}
\concept{module}
