% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_import.R
\name{mod_ui_import}
\alias{mod_ui_import}
\title{Import Spectra}
\usage{
mod_ui_import(id)
}
\arguments{
\item{id}{TODO.}
}
\description{
Import Spectra
}
\seealso{
Other module: 
\code{\link{mod_ui_about}()},
\code{\link{mod_ui_energy}()},
\code{\link{mod_ui_settings}()}
}
\concept{module}
