% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_settings.R
\name{mod_ui_settings}
\alias{mod_ui_settings}
\title{Settings}
\usage{
mod_ui_settings(id)
}
\arguments{
\item{id}{TODO.}
}
\description{
Settings
}
\seealso{
Other module: 
\code{\link{mod_ui_about}()},
\code{\link{mod_ui_energy}()},
\code{\link{mod_ui_import}()}
}
\concept{module}
