% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dose_server.R
\name{module_dose_server}
\alias{module_dose_server}
\title{Dose Rate Server}
\usage{
module_dose_server(input, output, session, user_data, user_settings)
}
\arguments{
\item{input}{provided by \pkg{Shiny}.}

\item{output}{provided by \pkg{Shiny}.}

\item{session}{provided by \pkg{Shiny}.}

\item{user_data}{a [shiny::reactiveValues()] list with the
following elements: "`code`", "`names`" and "`raw`".}

\item{user_settings}{a [shiny::reactiveValues()] list.}
}
\description{
Dose Rate Server
}
\seealso{
[module_dose_ui()]

Other modules: 
\code{\link{module_about_ui}()},
\code{\link{module_dose_ui}()},
\code{\link{module_energy_server}()},
\code{\link{module_energy_ui}()},
\code{\link{module_import_server}()},
\code{\link{module_import_ui}()},
\code{\link{module_settings_server}()},
\code{\link{module_settings_ui}()}
}
\concept{modules}
