% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dose_ui.R
\name{module_dose_ui}
\alias{module_dose_ui}
\title{Dose Rate UI}
\usage{
module_dose_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Dose Rate UI
}
\seealso{
\link{module_dose_server}

Other module: 
\code{\link{module_about_ui}()},
\code{\link{module_dose_server}()},
\code{\link{module_energy_server}()},
\code{\link{module_energy_ui}()},
\code{\link{module_import_server}()},
\code{\link{module_import_ui}()},
\code{\link{module_settings_server}()},
\code{\link{module_settings_ui}()}
}
\concept{module}
