% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_energy_ui.R
\name{module_energy_ui}
\alias{module_energy_ui}
\title{Energy Calibration UI}
\usage{
module_energy_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Energy Calibration UI
}
\seealso{
\link{module_import_server}

Other module: 
\code{\link{module_about_ui}()},
\code{\link{module_dose_server}()},
\code{\link{module_dose_ui}()},
\code{\link{module_energy_server}()},
\code{\link{module_import_server}()},
\code{\link{module_import_ui}()},
\code{\link{module_settings_server}()},
\code{\link{module_settings_ui}()}
}
\concept{module}
