% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_import_server.R
\name{module_import_server}
\alias{module_import_server}
\title{Import Server}
\usage{
module_import_server(input, output, session, user_data, user_settings)
}
\arguments{
\item{input}{provided by \pkg{Shiny}.}

\item{output}{provided by \pkg{Shiny}.}

\item{session}{provided by \pkg{Shiny}.}

\item{user_data}{a \code{\link[shiny]{reactiveValues}} list with the
following elements: "\code{spectra}", "\code{names}" and "\code{raw}".}

\item{user_settings}{a \code{\link[shiny]{reactiveValues}} list.}
}
\description{
Import Server
}
\seealso{
\link{module_import_ui}

Other module: 
\code{\link{module_about_ui}()},
\code{\link{module_dose_server}()},
\code{\link{module_dose_ui}()},
\code{\link{module_energy_server}()},
\code{\link{module_energy_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_settings_server}()},
\code{\link{module_settings_ui}()}
}
\concept{module}
