% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_import_ui.R
\name{module_import_ui}
\alias{module_import_ui}
\title{Import UI}
\usage{
module_import_ui(id)
}
\arguments{
\item{id}{A [`character`] vector to be used for the namespace.}
}
\description{
Import UI
}
\seealso{
[module_import_server()]

Other modules: 
\code{\link{module_about_ui}()},
\code{\link{module_dose_server}()},
\code{\link{module_dose_ui}()},
\code{\link{module_energy_server}()},
\code{\link{module_energy_ui}()},
\code{\link{module_import_server}()},
\code{\link{module_settings_server}()},
\code{\link{module_settings_ui}()}
}
\concept{modules}
