% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_settings_server.R
\name{module_settings_server}
\alias{module_settings_server}
\title{Settings Server}
\usage{
module_settings_server(input, output, session, user_settings)
}
\arguments{
\item{input}{provided by \pkg{Shiny}.}

\item{output}{provided by \pkg{Shiny}.}

\item{session}{provided by \pkg{Shiny}.}

\item{user_settings}{a \code{\link[shiny]{reactiveValues}} list.}
}
\description{
Settings Server
}
\seealso{
\link{module_settings_ui}

Other module: 
\code{\link{module_about_ui}()},
\code{\link{module_dose_server}()},
\code{\link{module_dose_ui}()},
\code{\link{module_energy_server}()},
\code{\link{module_energy_ui}()},
\code{\link{module_import_server}()},
\code{\link{module_import_ui}()},
\code{\link{module_settings_ui}()}
}
\concept{module}
