% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_app}
\alias{run_app}
\alias{launch_app}
\title{Run a Shiny App}
\usage{
run_app(app = c("doserate", "calibration"), browser = TRUE, display = "auto")

launch_app(
  app = c("doserate", "calibration"),
  browser = TRUE,
  display = "auto"
)
}
\arguments{
\item{app}{A [`character`] string specifying the Shiny application
to run. It must be one of "`doserate`" or "`calibration`"
(see details). Any unambiguous substring can be given.}

\item{browser}{A [`logical`] scalar: should the app be run in
the browser?}

\item{display}{A [`character`] string specifying the mode in which
to display the application (see [shiny::runApp()]).}
}
\value{
A \pkg{shiny} application object.
}
\description{
A wrapper for [shiny::shinyAppDir()].
}
\details{
\tabular{ll}{
  **Application name** \tab  **Keyword** \cr
  Dose rate estimation \tab `doserate` \cr
  Calibration curve builder \tab `calibration` \cr
 }
}
\note{
Usefull presets for the energy scale calibration:
 \tabular{ll}{
  Channel \tab Energy (keV) \cr
  76 \tab 238.63 \cr
  465 \tab 1460.82 \cr
  830 \tab 2614.51 \cr
 }
}
\examples{
\dontrun{
run_app("doserate")
run_app("calibration")
}
}
\author{
N. Frerebeau
}
\concept{shiny}
