% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaShiny-package.R
\docType{package}
\name{gammaShiny-package}
\alias{gammaShiny-package}
\alias{gammaShiny}
\title{gammaShiny: 'shiny' Applications for the R Package 'gamma'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of 'shiny' applications for the R package 'gamma'.
}
\details{
\tabular{ll}{
 **Package:** \tab gammaShiny \cr
 **Type:** \tab Package \cr
 **Version:** \tab 0.2.0 \cr
 **License:** \tab GPL-3 \cr
 **Zenodo:** \tab \doi{10.5281/zenodo.4139005}
}
}
\note{
This work received a state financial support managed by the Agence Nationale
de la Recherche (France) through the program *Investissements d'avenir*
(ref. ANR-10-LABX-52 and 11-IDEX-0001).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/crp2a/gammaShiny}
  \item Report bugs at \url{https://github.com/crp2a/gammaShiny/issues}
}

}
\author{
**Full list of authors and contributors** (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab *Université Bordeaux Montaigne, France* \cr
 Sebastian Kreutzer \tab *Universität Heidelberg, Germany* \cr
 Brice Lebrun \tab *Université Bordeaux Montaigne, France*
}

**Package maintainer**

Nicolas Frerebeau\cr
\email{services-archeosciences@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
