% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_app}
\alias{run_app}
\title{Run a Shiny App}
\usage{
run_app(app = c("doserate", "calibration"), browser = TRUE)
}
\arguments{
\item{app}{A \code{\link{character}} string specifying the Shiny application
to run. It must be one of "\code{doserate}" or "\code{calibration}"
(see details). Any unambiguous substring can be given.}

\item{browser}{A \code{\link{logical}} scalar: should the app be run in
the browser?}
}
\value{
A \pkg{shiny} application object.
}
\description{
A wrapper for \code{\link[shiny]{shinyAppDir}}.
}
\details{
\tabular{ll}{
  **Application name** \tab  **Keyword** \cr
  Dose rate estimation \tab \code{doserate} \cr
  Calibration curve builder \tab \code{calibration} \cr
 }
}
\examples{
\dontrun{
run_app("doserate")
run_app("calibration")
}
}
\author{
N. Frerebeau
}
\concept{shiny}
